requirements.txt
gdpc==5.0.2
pyglm
pyglm-typing
numpy
scipy
more-itertools
